Move logic from main to package apiutil
