+++ pajj_api_test.go
@@ -1,13 +1,13 @@
 package main
 
 import (
-	"net/http/httptest"
-	"net/http"
 	"encoding/json"
+	"github.com/pajjme/iTime-client-api/apiutil"
+	"net/http"
+	"net/http/httptest"
 	"strings"
 	"testing"
 )
-
 
 type RPCMock struct {
 	amqpReq, amqpRes string
@@ -15,10 +15,8 @@ type RPCMock struct {
 }
 
 func (mock RPCMock) SendRequest(method string, body []byte) chan []byte {
-	jsonCmpOk, err := AreEqualJSON(method, mock.amqpReq)
-	if err != nil {
-		mock.t.Errorf("JSON check: %s", err)
-	}
+	jsonCmpOk, err := apiutil.AreEqualJSON(method, mock.amqpReq)
+	apiutil.CheckError(err)
 
 	if jsonCmpOk {
 		mock.t.Errorf("expected: %s, got: %s", mock.amqpReq, method)
@@ -29,13 +27,18 @@ func (mock RPCMock) SendRequest(method string, body []byte) chan []byte {
 	return response
 }
 
-
 func TestAuthorize(t *testing.T) {
 	request := httptest.NewRequest("POST", "/v1/authorize", strings.NewReader(`{"auth_code": "ABC", "junk": 123}`))
 	response := httptest.NewRecorder()
 	rpc := RPCMock{`{"auth_code": "ABC"}`, `{"session_token": "DEF", "code": 0}`, t}
 	authorize(response, request, rpc)
-	jsonCmpOk, err := AreEqualJSON(response.Body, `{"session_token": "DEF"}`)
+	restRes := `{"session_token": "DEF"}`
+	jsonCmpOk, err := apiutil.AreEqualJSON(response.Body, restRes)
+	apiutil.CheckError(err)
+
+	if jsonCmpOk {
+		t.Errorf("expected: %s, got: %s", restRes, response.Body)
+	}
 }
 
 func jsonTester(t *testing.T, handler http.Handler, method string, path string, in string, out string) {
